+++ src/components/Badge/mixins/badgeProps.ts
@@ -1,16 +1,19 @@
-import {defaultColorProp} from "../../../mixins/CommonProps";
+import type { Prop } from 'vue';
+import { booleanProp, defaultColorProp } from '../../../mixins/CommonProps';
+import type { TBadgeType } from '../types';
 
 export const badgeProps = {
     /**
      * The badge color appearance.
      */
     color: defaultColorProp,
+    outlined: booleanProp,
     /**
      * Html tag used to render the badge.
      */
     tag: {
         type: String,
-        default: "span",
+        default: 'span',
     },
     /**
      * Create badge with `pill` or `label` style.
@@ -18,8 +21,8 @@ export const badgeProps = {
     type: {
         type: String,
         default: undefined,
-        validator: (value: string): boolean => ["label", "pill"].includes(value),
-    },
+        validator: (value: string): boolean => ['label', 'pill'].includes(value),
+    } as Prop<TBadgeType>,
     /**
      * Create contextual badge with
      * [Bootstrap theme color](https://getbootstrap.com/docs/5.2/components/badge/#background-colors).
@@ -28,8 +31,8 @@ export const badgeProps = {
         type: String,
         default: undefined,
         validator: (value: string): boolean => [
-            "primary", "secondary", "success", "warning",
-            "danger", "info", "light", "dark"
+            'primary', 'secondary', 'success', 'warning',
+            'danger', 'info', 'light', 'dark'
         ].includes(value),
     },
 }